refactor: base callback module
