+++ wandb_addons/diffusers/callbacks/base/__init__.py
@@ -1,0 +1,10 @@
+from .base_callback import BaseDiffusersCallback
+from .base_image_to_image_callback import BaseImage2ImageCallback
+from .base_multi_pipeline_callback import BaseMultiPipelineCallback
+
+
+__all__ = [
+    "BaseDiffusersCallback",
+    "BaseImage2ImageCallback",
+    "BaseMultiPipelineCallback",
+]
